+++ frontend/src/components/AnnulmentTransactionsTable/AnnulmentTransactionsTable.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import AnullmentTransactionsTable from './AnullmentTransactionsTable';
+import AnnulmentTransactionsTable from './AnnulmentTransactionsTable';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<AnullmentTransactionsTable userLevel={0}/>, div);
+  ReactDOM.render(<AnnulmentTransactionsTable userLevel={0}/>, div);
   ReactDOM.unmountComponentAtNode(div);
 });
