+++ frontend/src/components/index.js
@@ -1,5 +1,5 @@
 export { default as TransactionOverviewTable } from './TransactionOverviewTable/TransactionOverviewTable'
-export { default as AnullmentTransactionsTable } from './AnullmentTransactionsTable/AnullmentTransactionsTable'
+export { default as AnnulmentTransactionsTable } from './AnnulmentTransactionsTable/AnnulmentTransactionsTable'
 export { default as UserOverviewTable } from './UserOverviewTable/UserOverviewTable'
 export { default as LoginButton } from './LoginButton/LoginButton'
 export { default as WebFrame } from './WebFrame/WebFrame'
